+++ tests/planning/methods/ShortcutPath.py
