+++ tests/planning/test_MacSmoother.py
@@ -1,0 +1,9 @@
+from methods import SmoothTrajectoryTest
+from planning_helpers import BasePlannerTest
+from prpy.planning.mac_smoother import MacSmoother
+from unittest import TestCase
+
+class MacSmootherTests(BasePlannerTest,
+                       SmoothTrajectoryTest,
+                       TestCase):
+    planner_factory = MacSmoother
