+++ tests/planning/test_OMPL.py
@@ -1,0 +1,10 @@
+from methods import PlanToConfigurationTest
+from planning_helpers import BasePlannerTest
+from prpy.planning.ompl import OMPLPlanner
+from unittest import TestCase
+
+
+class OMPLPlannerTests(BasePlannerTest,
+                       PlanToConfigurationTest,
+                       TestCase):
+    planner_factory = OMPLPlanner
