+++ tests/planning/test_OMPLSimplifier.py
@@ -1,0 +1,10 @@
+from methods import ShortcutPathTest
+from planning_helpers import BasePlannerTest
+from prpy.planning.ompl import OMPLSimplifier
+from unittest import TestCase
+
+
+class OMPLSimplifierTests(BasePlannerTest,
+                          ShortcutPathTest,
+                          TestCase):
+    planner_factory = OMPLSimplifier
