+++ tests/planning/test_ParabolicRetimer.py
@@ -1,0 +1,9 @@
+from methods import RetimeTrajectoryTest
+from planning_helpers import BasePlannerTest
+from prpy.planning.retimer import ParabolicRetimer
+from unittest import TestCase
+
+class ParabolicRetimerTests(BasePlannerTest,
+                            RetimeTrajectoryTest,
+                            TestCase):
+    planner_factory = ParabolicRetimer
